Fix DF11 interrogator ID truncation accepting corrupted all-call replies
The DF=11 CRC syndrome (parity XOR interrogator code) was stored in a
uint16_t, so any corrupted all-call reply whose 24-bit syndrome only had
bits 16-23 set was treated as an II=0 reply and marked valid. That raises
the false-accept rate for noise frames that look like DF=11 by 256x, and
every accepted DF=11 creates an aircraft in the dictionary.

Keep the full 24-bit syndrome.

+++ firmware/common/adsb/mode_s_packet.cpp
@@ -239,7 +239,9 @@ void DecodedModeSPacket::ConstructModeSPacket() {
         case kDownlinkFormatAllCallReply:  // DF = 11
         {
             icao_address = Get24BitsFromWordBuffer(8, raw.buffer);
-            uint16_t interrogator_id = parity_value ^ calculated_checksum;
+            // The interrogator ID is the full 24-bit syndrome. Truncating it would accept corrupted packets whose
+            // syndrome happens to only have bits set above the truncation point.
+            uint32_t interrogator_id = parity_value ^ calculated_checksum;
             if (interrogator_id == 0) {
                 // Reply to a spontaneous acquisition squitter.
                 is_valid = true;
